+++ tests/test_settings.py
@@ -4,7 +4,7 @@ from __future__ import annotations
 
 from xloft import AliasDict
 
-from scruby_full_text import FullTextSettings
+from scruby_fts import FTSConfig
 
 TEST_MORPHOLOGY = AliasDict(
     ({"Arabic", "ar"}, "libstemmer_ar"),
@@ -44,7 +44,7 @@ class TestPositive:
     def test_supported_languages(self) -> None:
         """Test supported languages."""
         for lang, morphology in TEST_MORPHOLOGY.items():
-            morphology_by_lang_name = FullTextSettings.morphology.get(lang[0])
-            morphology_by_lang_code = FullTextSettings.morphology.get(lang[1])
+            morphology_by_lang_name = FTSConfig.morphology.get(lang[0])
+            morphology_by_lang_code = FTSConfig.morphology.get(lang[1])
             assert morphology_by_lang_name == morphology
             assert morphology_by_lang_code == morphology