+++ CHANGELOG.md
@@ -24,7 +24,7 @@
 
 #### v0.3.0 2026-01-20
 
-- Moved all parameters in the settings module to the `FullTextSettings` class.
+- Moved all parameters in the settings module to the `FTSConfig` class.
 - Upgraded dependencies.
 - Updated tests.
 - Updated README.md